requirements.txt
streamlit==1.19.0
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.1.3
Pillow==9.4.0